Adding --junit option to show subcommand
  1)  Moved junit creation code from elfws/subcommand/report.py to elfws/report/junit.py
  2)  Added --junit option to show subcommand
