+++ elfws/subcommand/show.py
@@ -2,9 +2,10 @@
 import re
 import sys
 
-from elfws import utils
 from elfws import display
+from elfws.report import junit
 from elfws import suppression_list
+from elfws import utils
 
 
 def show(cla):
@@ -16,3 +17,7 @@ def show(cla):
     oWarnList = utils.create_warning_list(lLogFile, cla.log_file)
 
     display.results(cla.log_file, '<None>', oSupList, oWarnList)
+
+    if cla.junit:
+        lJUnitFile = junit.generate_junit_xml_file(cla, oWarnList, oSupList)
+        utils.write_file(cla.junit, lJUnitFile)
